Try to read/write on all revents
stdout.revents can be set to POLLERR when the process plugged to stdout
is killed.

If this happens, we need to set out_closed.

This commits removes the filter "&(POLLOUT|POLLHUP)" to always try to
read or write: if there is an error, it will fail and properly set
out_closed as expected.